# src/openapi.json
{
  "openapi": "2.8.0",
  "info": {
    "title": "XTE Apis",
    "author": "D2:业",
    "version": "v1.0.0",
    "description": "Simple and easy to use API."
  },
  "servers": [
    {
      "url": "https://api.xte.web.id"
    }
  ],
...
  "paths": {
...
    "/v1/tools/upscale": {
      "get": {
        "summary": "AI Image Upscaler",
        "description": "Meningkatkan kualitas gambar (HDR) dan resolusi secara otomatis. Output langsung berupa file gambar PNG.",
        "tags": ["Tools"],
        "parameters": [
          { "name": "url", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "scale", "in": "query", "schema": { "type": "integer", "enum": [2, 4, 6, 8, 16], "default": 2 } },
          { "name": "anime", "in": "query", "schema": { "type": "string", "enum": ["true", "false"], "default": "false" } }
        ],
        "responses": { "200": { "description": "Image PNG" } }
      }
    },
    "/v1/tools/kbbi": {
      "get": {
        "summary": "KBBI Dictionary",
        "description": "Mencari definisi kata baku dalam Kamus Besar Bahasa Indonesia (KBBI).",
        "tags": ["Tools"],
        "parameters": [{ "name": "kata", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success" } }
      }
    },
...
  }
}